Properly handle cases when server switch did not include configuration phase.
